refactor: Move path logic to from `utils` to `path_utils` in polars-io
