+++ crates/polars-io/src/lib.rs
@@ -21,6 +21,7 @@ mod options;
 pub mod parquet;
 #[cfg(feature = "parquet")]
 pub mod partition;
+pub mod path_utils;
 #[cfg(feature = "async")]
 pub mod pl_async;
 pub mod predicates;
@@ -31,6 +32,7 @@ pub mod utils;
 #[cfg(feature = "cloud")]
 pub use cloud::glob as async_glob;
 pub use options::*;
+pub use path_utils::*;
 pub use shared::*;
 
 pub mod hive;
