+++ crates/polars-io/src/prelude.rs
@@ -11,5 +11,6 @@ pub use crate::ndjson::core::*;
 pub use crate::parquet::{metadata::*, read::*, write::*};
 #[cfg(feature = "parquet")]
 pub use crate::partition::write_partitioned_dataset;
+pub use crate::path_utils::*;
 pub use crate::shared::{SerReader, SerWriter};
 pub use crate::utils::*;
