+++ crates/polars-io/src/utils/mod.rs
@@ -1,5 +1,3 @@
 mod other;
-mod path;
 
 pub use other::*;
-pub use path::*;
