+++ crates/polars-mem-engine/src/executors/scan/parquet.rs
@@ -7,7 +7,7 @@ use polars_core::config::{get_file_prefetch_size, verbose};
 use polars_core::utils::accumulate_dataframes_vertical;
 use polars_io::cloud::CloudOptions;
 use polars_io::parquet::metadata::FileMetaDataRef;
-use polars_io::utils::is_cloud_url;
+use polars_io::path_utils::is_cloud_url;
 use polars_io::RowIndex;
 
 use super::*;
