+++ crates/polars-pipe/src/executors/sources/csv.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 
 use polars_core::{config, POOL};
 use polars_io::csv::read::{BatchedCsvReader, CsvReadOptions, CsvReader};
-use polars_io::utils::is_cloud_url;
+use polars_io::path_utils::is_cloud_url;
 use polars_plan::global::_set_n_rows_for_scan;
 use polars_plan::prelude::FileScanOptions;
 use polars_utils::iter::EnumerateIdxTrait;
