+++ crates/polars-plan/src/plans/conversion/scans.rs
@@ -1,10 +1,10 @@
 use std::path::PathBuf;
 
 use either::Either;
+use polars_io::path_utils::is_cloud_url;
 #[cfg(feature = "cloud")]
 use polars_io::pl_async::get_runtime;
 use polars_io::prelude::*;
-use polars_io::utils::is_cloud_url;
 use polars_io::RowIndex;
 
 use super::*;
